Practica01-Temperatura/requirements.txt
tensorflow-macos
tensorflow-metal
pandas
numpy
seaborn
matplotlib